+++ packages/core/src/components/screens/release-notes/ReleaseNotesScreen.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import {
     Accordion,
     CircularProgress,
@@ -6,44 +6,14 @@ import {
     Typography,
     useStyles,
 } from "@equinor/mad-components";
-import { fetchAllReleaseNotes } from "./fetchReleaseNotes";
 import { ChangeLog, Release } from "./ChangeLog";
-import { useEnvironment, useServicePortalName } from "../../../store/mad-config";
 import { getShortDate } from "../../../utils/dateUtils";
 import { ScrollView, View } from "react-native";
-import * as mockData from "../../../static/mock-data/whats-new.json";
-import {useDemoMode} from "../../../store/demo-mode";
+import { useReleaseNotes } from "../../../hooks/UseReleaseNotes";
 
 export const ReleaseNotesScreen = () => {
     const styles = useStyles(releaseNoteStyles);
-    const demoMode = useDemoMode();
-    const environment = useEnvironment();
-    const servicePortalName = useServicePortalName();
-
-    const [releaseNotes, setReleaseNotes] = useState<Release[] | null>(null);
-    const [error, setError] = useState("");
-    const [isFetching, setIsFetching] = useState(true);
-
-    useEffect(() => {
-        if (demoMode.isEnabled) {
-            setReleaseNotes([mockData]);
-            setIsFetching(false);
-        } else {
-            fetchAllReleaseNotes(environment, servicePortalName)
-                .then(setReleaseNotes)
-                .catch((error: Error) => setError(error.message))
-                .finally(() => setIsFetching(false));
-        }
-
-    }, [demoMode.isEnabled, environment, servicePortalName]);
-
-    const sortReleaseNotes = (releaseNoteA: Release, releaseNoteB: Release) => {
-        return releaseNoteA.version < releaseNoteB.version
-            ? releaseNoteB.version < releaseNoteA.version
-                ? -1
-                : 1
-            : 0;
-    };
+    const { releaseNotes, isFetching, error } = useReleaseNotes();
 
     if (isFetching) {
         return (
@@ -53,8 +23,6 @@ export const ReleaseNotesScreen = () => {
         );
     }
     if (releaseNotes && releaseNotes.length > 0) {
-        releaseNotes.sort(sortReleaseNotes);
-
         return (
             <ScrollView style={styles.container}>
                 <Accordion>
@@ -66,7 +34,7 @@ export const ReleaseNotesScreen = () => {
                             )}`}
                             chevronPosition={"right"}
                         >
-                            <ChangeLog release={releaseNote} showTitleAndVersion={false}/>
+                            <ChangeLog release={releaseNote} />
                         </Accordion.Item>
                     ))}
                 </Accordion>
