+++ packages/core/src/components/screens/release-notes/WhatsNewScreen.tsx
@@ -3,11 +3,19 @@ import { ChangeLog, Release } from "./ChangeLog";
 import * as mockData from "../../../static/mock-data/whats-new.json";
 import { useAppVersion, useEnvironment, useServicePortalName } from "../../../store/mad-config";
 import { useReleaseNotesVersion } from "../../../store/release-notes";
-import {Button, CircularProgress, EDSStyleSheet, useStyles} from "@equinor/mad-components";
-import { useCoreStackNavigation } from "../../../hooks/useCoreStackNavigation";
+import {
+    Button,
+    Cell,
+    CircularProgress,
+    EDSStyleSheet,
+    Typography,
+    useStyles,
+} from "@equinor/mad-components";
+import { useCoreStackNavigation } from "../../../hooks";
 import { useDemoMode } from "../../../store/demo-mode";
 import { fetchReleaseNotes } from "./fetchReleaseNotes";
-import { View } from "react-native";
+import { ScrollView, View } from "react-native";
+import { getShortDate } from "../../../utils/dateUtils";
 
 /**
  * This screen will display the latest releasenotes
@@ -54,7 +62,15 @@ export const WhatsNewScreen = () => {
 
     return (
         <View style={styles.container}>
-            <ChangeLog release={release} />
+            <ScrollView>
+                <Cell style={styles.scrollContainer}>
+                    <Typography style={styles.versionHeader}>{release.version}</Typography>
+                    <Typography style={styles.subtitleHeader}>
+                        {getShortDate(new Date(release.releaseDate))}
+                    </Typography>
+                    <ChangeLog release={release} />
+                </Cell>
+            </ScrollView>
             <View style={styles.footer}>
                 <Button
                     title="OK"
@@ -69,18 +85,27 @@ export const WhatsNewScreen = () => {
     );
 };
 
-const whatsNewStyles = EDSStyleSheet.create((theme) => ({
+const whatsNewStyles = EDSStyleSheet.create(theme => ({
     spinnerContainer: {
         display: "flex",
         height: "100%",
         justifyContent: "center",
-        alignItems: "center"
+        alignItems: "center",
     },
     container: {
         display: "flex",
         paddingTop: theme.geometry.dimension.cell.minHeight,
-        height: "100%",
         justifyContent: "space-between",
+    },
+    scrollContainer: {
+        justifyContent: "center",
+        borderTopWidth: 0,
+        borderBottomWidth: 0,
+    },
+    versionHeader: theme.typography.basic.h1,
+    subtitleHeader: {
+        ...theme.typography.basic.h4,
+        marginVertical: theme.spacing.container.paddingVertical,
     },
     footer: {
         alignItems: "flex-end",
